FarhanNurzi/WALL-1657/Link the Wallet carousel with real data in Responsive (#9895)
* chore: use hook to get wallet list in mobile carousel

* chore: add components/index.ts
