+++ packages/wallets/src/components/AccountsList/AccountsList.scss
@@ -2,4 +2,5 @@
     height: 60vh;
     width: 100vw;
     margin-top: 2rem;
+    background-color: #add8e6;
 }
