+++ packages/wallets/src/components/AccountsList/AccountsList.tsx
@@ -1,18 +1,12 @@
 import React from 'react';
+import { useWalletAccountsList } from '@deriv/api';
 
 type TAccountsListProps = {
-    data: {
-        text: string;
-        background: string;
-    };
+    data: ReturnType<typeof useWalletAccountsList>['data'][number];
 };
 
 const AccountsList = ({ data }: TAccountsListProps) => {
-    return (
-        <div className='wallets-accounts-list' style={{ backgroundColor: data.background }}>
-            AccountsList
-        </div>
-    );
+    return <div className='wallets-accounts-list'>{data.loginid}</div>;
 };
 
 export default AccountsList;
