--- packages/wallets/src/components/WalletCarousel/WalletsCarousel.scss
@@ -1,19 +1,0 @@
-.wallets-carousel {
-    background-color: #f2f3f4;
-    width: 100vw;
-    padding: 2rem;
-
-    &__container {
-        height: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        gap: 2.4rem;
-    }
-}
-
-.wallets-card {
-    width: 100%;
-    height: 13rem;
-    border-radius: 10%;
-}
