+++ packages/wallets/src/components/WalletsCarousel/WalletsCarousel.scss
@@ -1,0 +1,53 @@
+.wallets-carousel {
+    background-color: #f2f3f4;
+    width: 100vw;
+    padding: 2rem;
+    overflow: hidden;
+
+    &__container {
+        height: 100%;
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        gap: 2.4rem;
+    }
+}
+
+.wallets-card {
+    width: 100%;
+    height: 13rem;
+    background-color: #add8e6;
+
+    &__data {
+        display: flex;
+        justify-content: space-between;
+        width: 100%;
+        height: 100%;
+        padding: 1.6rem;
+
+        &__details {
+            display: flex;
+            flex-direction: column;
+            justify-content: space-between;
+
+            &-balance {
+                display: flex;
+                flex-direction: column;
+                gap: 1rem;
+
+                & > h3 {
+                    font-size: 1.6rem;
+                    font-weight: bold;
+                }
+            }
+        }
+
+        &__landing-company {
+            font-size: 1rem;
+            font-weight: bold;
+            padding: 0.5rem;
+            height: max-content;
+            border: 1px solid #000000;
+        }
+    }
+}
