+++ packages/wallets/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './AccountsList';
+export * from './WalletsCarousel';
+export * from './WalletList';
